🐛 bug: use client 파일 분리
